# Swap minimist for getopts for more perf gains!