Hi @morishitter! This PR swaps minimist for [getopts](https://github.com/jorgebucaran/getopts) for more perf gains (full benchmarks [here](https://github.com/jorgebucaran/getopts/tree/master/bench)).

<pre>
mri × 374,999 ops/sec
yargs × 32,370 ops/sec
<b>getopts × 1,495,165 ops/sec</b>
minimist × 283,031 ops/sec
</pre>

Getopts has essentially the same API as minimist, so this is effectively a "drop-in" replacement. I think minimist is a great package, but also too difficult to contribute to (34 open issues and 26 pull requests, but no updates since 2015).

Let me know if this looks good to you. Cheers!